requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.8
django-cleanup==8.1.0
django-image-optimizer==1.0.3
idna==3.6
pillow==10.2.0
python-decouple==3.8
python-resize-image==1.1.20
requests==2.31.0
sqlparse==0.4.4
tinify==1.6.0
tzdata==2024.1
urllib3==2.2.1
